[Feature] Update data type and column about point [Issue no] #136

+++ src/main/java/com/impacus/maketplace/entity/point/greenLablePoint/GreenLabelPointAllocation.java
@@ -29,7 +29,7 @@ public class GreenLabelPointAllocation extends BaseEntity {
     private PointStatus pointStatus;
 
     @Column(nullable = false)
-    private int remainPoint;
+    private Long remainPoint;
 
     @Column(nullable = false)
     private LocalDateTime expiredAt;
