+++ src/main/java/com/impacus/maketplace/entity/point/greenLablePoint/GreenLabelPointHistory.java
@@ -25,5 +25,5 @@ public class GreenLabelPointHistory extends BaseEntity {
     private PointStatus pointStatus;
 
     @Column(nullable = false)
-    private int tradeAmount;
+    private Long tradeAmount;
 }
